feat: 添加客户端内自动更新、被动吊射观察、设置面板、连接状态HUD等功能
- StartupUpdateChecker: 客户端内检测更新+下载+原子替换重启，含轮播相册
- LobShotService: 支持比赛模式被动观察（Shift进入/Escape退出，不发送指令）
- SettingsPanel/GameParamsConfig: 运行时配置面板与JSON持久化
- ConnectionStatusHUD/ExitConfirmPanel/QuickReferencePanel: 新增HUD组件
- BuildScript+publish_http.sh+launch.sh/bat: 自动构建+发布+启动器工具链
- 文档: 修正视觉吊射方案、自定义数据传输协议等多处不一致
- 通信协议文档升级到 V1.3.0

+++ Assets/Scripts/UI/ViewModels/RuneStatusSyncViewModel.cs
@@ -6,11 +6,12 @@ namespace UI.ViewModels
     {
         private uint runeStatus;
         private uint activatedArms;
-        private uint averageRings;
+        private float averageRings;
 
         public uint RuneStatus { get => runeStatus; set { if (runeStatus != value) { runeStatus = value; OnPropertyChanged(); } } }
         public uint ActivatedArms { get => activatedArms; set { if (activatedArms != value) { activatedArms = value; OnPropertyChanged(); } } }
-        public uint AverageRings { get => averageRings; set { if (averageRings != value) { averageRings = value; OnPropertyChanged(); } } }
+        /// <summary>V1.3.0: 从 uint32(总环数) 改为 float(平均环数)，仅大能量机关正在激活时有效</summary>
+        public float AverageRings { get => averageRings; set { if (averageRings != value) { averageRings = value; OnPropertyChanged(); } } }
 
         protected override void UpdateFrom(RuneStatusSync msg)
         {
